requirements.txt
aiogram==2.25.1
openai
matplotlib
fpdf
pydrive
SpeechRecognition
notion-client
aioschedule
python-dotenv